pyproject.toml
[tool.poetry]
name = "kernel-module-exporter"
version = "0.1.0"
description = "Expose metrics for kernel module"
authors = ["Sylvain Chateau <git@sylvain-chateau.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
prometheus-client = "^0.20.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
